Moved printing from the SolveClass back to Main, so all methods can use the same functions for printing.
This should simplify the ability to write tests for these functions, since they just return a linked list instead of a pre-formatted string to the console.
It also just makes the code cleaner in general, since we don't have to deal with string concatenation and printing in the solver class, potentially causing issues in the future.

Main:
- SolveClass functions now get given a full LabyrinthClass object
- Added a simple printing algorithm, so all prints are handled by Main

LabyrinthClass:
- Added the Solved boolean flag, for easy checking if a Maze has been solved.
- Added the Path LinkedList variable, so we can store and pass the taken path to other functions when needed.

SolveClass:
- Removed all printing logic.
- Refactored code to support Recieving a full LabyrinthClass object.
- Added Solved setters to the functions.