Fixed issue with RealWorldSolve method reporting a messed up/wrong path
The RealWorldSolve path is always written in to know where the solver has been for the final output, once encountering a dead end the solver didn't correctly delete the path, either deleting the entire path (including after the intersection) or not deleting the path at all.

This should solve those issues.